Render with a list of phrases instead of a phrase
For the tokenised phrases, this doesn't matter. The shape of a tokenised
phrase is identical to the shape of a list of tokenised phrases.

However, for the rendered phrases, there is a difference: a rendered
phrase contains an ID (and potentially other render-specific properties
down the line). The ID helps cache the render to speed up re-rendering
times when only part of the text has changed. However, the topmost level
will always have changed, so there's no point retaining an ID for it.

I'm not sure that this is necessary or if it even makes a difference.
Regardless, this is the first time I've used git commit without -m in
forever, and I'm not just gonna not commit this now that I've typed all
this out.