Clean up some stuff in the tokeniser
A user on Freenode prompted me to have a think about whether or not I
could replace the ugly stinky for...of loops in the tokeniser with
better declarative methods. The user recommended flatMap in particular
and I agreed to try it out, however, I wasn't able to implement it
without confusing myself. But I did manage to squeeze in a few
clean-ups.