Implement initial dot curves
The dot curves on letters that would have dots are now indicated with
debug lines. There are different implementations for full (p, f) and
unfull (s, d) blocks - full blocks have an entire circle, but unfull
blocks just have a start and end point with a curve drawn between them.
The angle on which to place the dots on a full block could possibly even
be random, although for the f block, it must not cross the word line.

The distance between the dot curve and the letter curve is a bit wack.
The current implementation is a naive, with the difference in radii just
being a ratio of the letter's radius, which is not ideal and is open to
unsightly differences between blocks.